Remove dependency on pytz, and make user worry about timezones.
We happen to know our servers will all have a UTC timezone, reported as 'Zulu' in the client handshake.

This change means that we can use this library out of the box on Ubuntu Xenial systems, provided we don't turn on compression. If we do, there is a stock python3-lz4 package, but not zstd.